+++ src/lib/bars/chart-bar-grouped.ts
@@ -1,6 +1,14 @@
 import { BaseType, select, Selection } from 'd3-selection';
 import { axisTickFindByIndex, axisTickHighlight } from '../axis';
-import { arrayIs2D, COLORS_CATEGORICAL, debug, nodeToString, siblingIndex } from '../core';
+import {
+  arrayFlat,
+  arrayIs2D,
+  COLORS_CATEGORICAL,
+  debug,
+  nodeToString,
+  siblingIndex,
+  siblingIndexSameClasses,
+} from '../core';
 import {
   chartCartesian,
   chartCartesianUpdateAxes,
@@ -15,20 +23,20 @@ import {
   legendItemFindByIndex,
   legendItemHighlight,
 } from '../legend';
-import { barFindByIndex, barHighlight, DataBar, JoinEvent } from './series-bar';
+import { barFindByCategory, barHighlight } from './series-bar';
 import {
-  barGroupedFindByGroupIndex,
+  barGroupedFindBySubcategory,
   DataBarGrouped,
   DataSeriesBarGrouped,
   dataSeriesBarGrouped,
   seriesBarGrouped,
 } from './series-bar-grouped';
 import { labelFind, labelFindByFilter, labelHighlight, seriesLabel } from './series-label';
-import { dataSeriesLabelBar } from './series-label-bar';
+import { DataSeriesLabelBar, dataSeriesLabelBar } from './series-label-bar';
 
 export interface DataChartBarGrouped extends DataSeriesBarGrouped, DataChartCartesian {
   legend: Partial<DataLegendSquares>;
-  subcategories: string[];
+  labels: Partial<DataSeriesLabelBar>;
 }
 
 export function dataChartBarGrouped(data: Partial<DataChartBarGrouped>): DataChartBarGrouped {
@@ -37,7 +45,7 @@ export function dataChartBarGrouped(data: Partial<DataChartBarGrouped>): DataCha
     ...seriesData,
     ...dataChartCartesian(data),
     legend: data.legend || {},
-    subcategories: data.subcategories || (seriesData.values[0] || []).map((d, i) => i.toString()),
+    labels: data.labels || {},
   };
 }
 
@@ -117,6 +125,7 @@ export function chartBarGroupedDataChange<
       } = chartData,
       s = select<GElement, Datum>(g[i]),
       barSeries = s.selectAll('.series-bar-grouped'),
+      labelSeries = s.selectAll<Element, DataSeriesLabelBar>('.series-label'),
       legend = s.selectAll<Element, DataLegendSquares>('.legend');
 
     barSeries.dispatch('datachange');
@@ -134,6 +143,14 @@ export function chartBarGroupedDataChange<
       )
     );
 
+    labelSeries.datum((d) =>
+      Object.assign<DataSeriesLabelBar, Partial<DataSeriesLabelBar>, Partial<DataSeriesLabelBar>>(
+        d,
+        { labels: arrayFlat(chartData.values).map((v) => v.toString()) },
+        chartData.labels
+      )
+    );
+
     xAxis.scale = categoryScale;
     yAxis.scale = valueScale;
     chartCartesianUpdateAxes(s);
@@ -146,14 +163,17 @@ export function chartBarGroupedDataChange<
 }
 
 export function chartBarGroupedHoverBar(
-  chart: Selection,
+  chart: Selection<Element, DataChartBarGrouped>,
   bar: Selection<SVGRectElement, DataBarGrouped>,
   hover: boolean
 ): void {
+  const chartD = chart.datum();
   bar.each((barD, i, g) => {
-    const labelS = labelFind(chart, barD.key),
-      tickS = axisTickFindByIndex(chart.selectAll('.axis-x'), barD.groupIndex),
-      legendItemS = legendItemFindByIndex(chart, barD.index);
+    const categoryIndex = chartD.categories.indexOf(barD.category),
+      subcategoryIndex = chartD.subcategories.indexOf(barD.subcategory),
+      labelS = labelFind(chart, barD.key),
+      tickS = axisTickFindByIndex(chart.selectAll('.axis-x'), categoryIndex),
+      legendItemS = legendItemFindByIndex(chart, subcategoryIndex);
 
     labelHighlight(labelS, hover);
     axisTickHighlight(tickS, hover);
@@ -162,14 +182,15 @@ export function chartBarGroupedHoverBar(
 }
 
 export function chartBarGroupedHoverLegendItem(
-  chart: Selection,
+  chart: Selection<Element, DataChartBarGrouped>,
   legendItem: Selection<Element, DataLegendItem>,
   hover: boolean
 ): void {
   const legendItemCount = chart.selectAll('.legend-item').size();
   legendItem.each((_, i, g) => {
     const legendItemI = siblingIndex(g[i], '.legend-item'),
-      barS = barFindByIndex<DataBarGrouped>(chart, legendItemI),
+      subcategory = chart.datum().subcategories[legendItemI],
+      barS = barGroupedFindBySubcategory(chart, subcategory),
       labelS = labelFindByFilter(
         chart.selectAll('.series-label'),
         (_, i) => i % legendItemCount === legendItemI
@@ -181,14 +202,15 @@ export function chartBarGroupedHoverLegendItem(
 }
 
 export function chartBarGroupedHoverAxisTick(
-  chart: Selection,
+  chart: Selection<Element, DataChartBarGrouped>,
   tick: Selection<Element>,
   hover: boolean
 ): void {
   const legendItemCount = chart.selectAll('.legend-item').size();
   tick.each((_, i, g) => {
-    const tickI = siblingIndex(g[i], '.tick'),
-      barS = barGroupedFindByGroupIndex(chart, tickI),
+    const tickI = siblingIndexSameClasses(g[i]),
+      category = chart.datum().categories[tickI],
+      barS = barFindByCategory(chart, category),
       labelS = labelFindByFilter(
         chart.selectAll('.series-label'),
         (_, i) => Math.floor(i / legendItemCount) === tickI