+++ src/lib/bars/index.ts
@@ -13,6 +13,7 @@ export * from './series-label';
 export * from './series-label-bar';
 export * from './chart-window-bar';
 export * from './chart-window-bar-grouped';
+export * from './chart-window-bar-stacked';
 
 // export * from './grouped-bars';
 // export * from './stacked-bars';