+++ src/lib/core/utility/array.ts
@@ -17,3 +17,18 @@ export function arrayIs(array: any): array is any[] {
 export function arrayIs2D(array: any): array is any[][] {
   return arrayIs(array) && array.every((e) => Array.isArray(e));
 }
+
+export type ValueOrArray<T> = T | ValueOrArray<T>[];
+export type NestedArray<T> = ValueOrArray<T>[];
+
+export function arrayFlat<T>(array: NestedArray<T>): T[] {
+  return arrayIs2D(array)
+    ? array.reduce((acc, val) => acc.concat(arrayIs2D(val) ? arrayFlat(val) : val), [])
+    : array.slice();
+}
+
+export function arrayPartition<T>(array: T[], size: number): T[][] {
+  const partitions: T[][] = [];
+  for (let i = 0; i < array.length; i += size) partitions.push(array.slice(i, i + size));
+  return partitions;
+}